Unify link manager code.
