+++ bundles/org.connectorio.addons.link/src/main/java/org/connectorio/addons/link/internal/AltChannelLinkManager.java
@@ -17,123 +17,28 @@
  */
 package org.connectorio.addons.link.internal;
 
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.CopyOnWriteArrayList;
-import java.util.concurrent.CopyOnWriteArraySet;
-import org.connectorio.addons.link.LinkListener;
 import org.connectorio.addons.link.LinkManager;
 import org.openhab.core.common.registry.Registry;
-import org.openhab.core.common.registry.RegistryChangeListener;
-import org.openhab.core.thing.ChannelUID;
-import org.openhab.core.thing.Thing;
 import org.openhab.core.thing.ThingRegistry;
-import org.openhab.core.thing.ThingUID;
 import org.openhab.core.thing.link.ItemChannelLink;
-import org.openhab.core.thing.link.ItemChannelLinkRegistry;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
-@Component
-public class AltChannelLinkManager implements LinkManager {
-
-  public static final String ITEM_CHANNEL_LINK_REGISTRY_FILTER = "(objectClass=org.openhab.core.thing.link.ItemChannelLinkRegistry)";
-
-  private final Map<Thing, List<LinkListener>> listeners = new ConcurrentHashMap<>();
-
-  private final Set<ChannelUID> linkedChannels = new CopyOnWriteArraySet<>();
-
-  private final Registry<ItemChannelLink, String> linkRegistry;
-  private final Registry<Thing, ThingUID> thingRegistry;
-
-  private final LinkRegistryListener linkRegistryListener;
-  private final ThingRegistryListener thingRegistryListener;
+/**
+ * This is an alternative channel link manager which rely on different inheritance tree of
+ * {@link org.openhab.core.thing.link.ItemChannelLinkRegistry}.
+ *
+ * Main point of this class is provisioning of {@link org.connectorio.addons.link.LinkManager}
+ * implementation which works with pure interfaces.
+ */
+@Component(service = LinkManager.class)
+public class AltChannelLinkManager extends BaseChannelLinkManager {
 
   @Activate
   public AltChannelLinkManager(@Reference(target = ITEM_CHANNEL_LINK_REGISTRY_FILTER) Registry<ItemChannelLink, String> linkRegistry,
     @Reference ThingRegistry thingRegistry) {
-    this.linkRegistry = linkRegistry;
-    this.thingRegistry = thingRegistry;
-
-    linkRegistryListener = new LinkRegistryListener();
-    thingRegistryListener = new ThingRegistryListener();
-    linkRegistry.addRegistryChangeListener(linkRegistryListener);
-    thingRegistry.addRegistryChangeListener(thingRegistryListener);
+    super(linkRegistry, thingRegistry);
   }
 
-  public void deactivate() {
-    linkRegistry.removeRegistryChangeListener(linkRegistryListener);
-    thingRegistry.removeRegistryChangeListener(thingRegistryListener);
-  }
-
-  @Override
-  public boolean isLinked(ChannelUID channelUID) {
-    return linkedChannels.contains(channelUID);
-  }
-
-  @Override
-  public boolean hasLinkedChannels(Thing thing) {
-    return linkedChannels.stream().anyMatch(channel -> thing.getUID().equals(channel.getThingUID()));
-  }
-
-  @Override
-  public void registerListener(Thing thing, LinkListener listener) {
-    listeners.computeIfAbsent(thing, (t) -> new CopyOnWriteArrayList<>())
-      .add(listener);
-  }
-
-  @Override
-  public void deregisterListener(Thing thing, LinkListener listener) {
-    List<LinkListener> linkListeners = listeners.get(thing);
-    if (linkListeners != null) {
-      linkListeners.remove(listener);
-      if (linkListeners.isEmpty()) {
-          listeners.remove(thing);
-      }
-    }
-  }
-
-  class LinkRegistryListener implements RegistryChangeListener<ItemChannelLink> {
-
-    @Override
-    public void added(ItemChannelLink element) {
-      linkedChannels.add(element.getLinkedUID());
-      Thing thing = thingRegistry.get(element.getLinkedUID().getThingUID());
-      if (thing != null && listeners.containsKey(thing)) {
-        listeners.get(thing).forEach(listener -> listener.linked(element.getLinkedUID()));
-      }
-    }
-
-    @Override
-    public void removed(ItemChannelLink element) {
-      Thing thing = thingRegistry.get(element.getLinkedUID().getThingUID());
-      if (thing != null && listeners.containsKey(thing)) {
-        listeners.get(thing).forEach(listener -> listener.unlinked(element.getLinkedUID()));
-      }
-      linkedChannels.remove(element.getLinkedUID());
-    }
-
-    @Override
-    public void updated(ItemChannelLink oldElement, ItemChannelLink element) {
-    }
-  }
-
-  class ThingRegistryListener implements RegistryChangeListener<Thing> {
-
-    @Override
-    public void added(Thing element) {
-    }
-
-    @Override
-    public void removed(Thing element) {
-
-    }
-
-    @Override
-    public void updated(Thing oldElement, Thing element) {
-    }
-  }
 }
